Move helper functions to separate folder
